refactor: member exception & success 이동 적용(#29)
